Moved glue.qt.widgets.data_viewer to glue.viewers.common.qt.data_viewer and glue.qt.widgets.mpl_widget to glue.viewers.common.qt.mpl_widget
