+++ glue/plugins/dendro_viewer/qt_widget.py
@@ -8,8 +8,8 @@ from glue.qt.mouse_mode import PickMode
 from glue.qt.qtutil import load_ui
 from glue.utils.qt.widget_properties import (ButtonProperty, CurrentComboProperty,
                                        connect_bool_button, connect_current_combo)
-from glue.qt.widgets.data_viewer import DataViewer
-from glue.qt.widgets.mpl_widget import MplWidget, defer_draw
+from glue.viewers.common.qt.data_viewer import DataViewer
+from glue.viewers.common.qt.mpl_widget import MplWidget, defer_draw
 from glue.utils import nonpartial
