+++ glue/qt/tests/test_toolbar.py
@@ -2,10 +2,11 @@
 
 from __future__ import absolute_import, division, print_function
 
+from glue.viewers.common.qt.mpl_widget import MplWidget
+
 from ..glue_toolbar import GlueToolbar
 from ..mouse_mode import MouseMode
 from ..qtutil import get_icon
-from ..widgets import MplWidget
 
 
 class MouseModeTest(MouseMode):
