+++ glue/qt/widgets/__init__.py
@@ -3,4 +3,3 @@ from __future__ import absolute_import, division, print_function
 from .custom_component_widget import CustomComponentWidget
 from .layer_tree_widget import LayerTreeWidget
 from .message_widget import MessageWidget
-from .mpl_widget import MplWidget
