+++ glue/viewers/common/qt/data_viewer.py
