+++ glue/viewers/common/qt/mpl_widget.py
