+++ glue/viewers/image/qt/viewer_widget.py
@@ -15,9 +15,9 @@ from glue.qt.mouse_mode import (RectangleMode, CircleMode, PolyMode,
 from glue.qt.qtutil import load_ui, get_icon
 from glue.utils.qt.widget_properties import CurrentComboProperty, ButtonProperty, connect_current_combo, _find_combo_data
 from glue.qt.widgets.data_slice_widget import DataSlice
-from glue.qt.widgets.data_viewer import DataViewer
+from glue.viewers.common.qt.data_viewer import DataViewer
 from glue.qt.widgets.glue_mdi_area import GlueMdiSubWindow
-from glue.qt.widgets.mpl_widget import MplWidget, defer_draw
+from glue.viewers.common.qt.mpl_widget import MplWidget, defer_draw
 from glue.utils import nonpartial, Pointer
 from glue.utils.qt import cmap2pixmap, update_combobox
